requirements.txt
# python=3.11.9
langchain 
beautifulsoup4 
black 
tiktoken 
openai 
pinecone-client 
unstructured 
nltk 
fastapi 
jinja2 
uvicorn 
streamlit 
streamlit-chat 
tqdm 
langchain-pinecone 
langchain-openai 
langchain-community
langchainhub
python-dotenv